Fix CSP issue by inlining CSS + logo and hide once the bundle is loaded
